+++ source/code/Makefile
@@ -45,7 +45,7 @@ endif
 
 BASE_LIBS += -lphysfs
 
-OFILES=main.o highscore.o ReadKeyboard.o joypad.o listFiles.o common.o stats.o CppSdlImageHolder.o Libs/NFont.o MenuSystem.o menudef.o puzzlehandler.o ${SAGO_O_FILES}
+OFILES=main.o highscore.o ReadKeyboard.o joypad.o listFiles.o common.o stats.o Libs/NFont.o Libs/SDL_FontCache.o MenuSystem.o menudef.o puzzlehandler.o ${SAGO_O_FILES}
 
 ifeq ($(CROSS),i686-pc-mingw32-)
 OFILES += winicon.res
