+++ source/code/main.cpp
@@ -92,87 +92,13 @@ http://blockattack.sf.net
 using namespace std;
 
 
-SDL_Surface* IMG_Load2(const char* path) {
-	if (!PHYSFS_exists(path)) {
-		cerr << "Error: File not in blockattack.data: " << path << endl;
-		return nullptr; //file doesn't exist
-	}
-
-	PHYSFS_file* myfile = PHYSFS_openRead(path);
-
-	// Get the lenght of the file
-	unsigned int m_size = PHYSFS_fileLength(myfile);
-
-	// Get the file data.
-	char* m_data = new char[m_size];
-
-	int length_read = PHYSFS_read (myfile, m_data, 1, m_size);
-
-	if (length_read != (int)m_size) {
-		delete [] m_data;
-		m_data = 0;
-		PHYSFS_close(myfile);
-		cerr << "Error: Curropt data file!" << endl;
-		return nullptr;
-	}
-
-	PHYSFS_close(myfile);
-
-// And this is how you load an image from a memory buffer with SDL
-	SDL_RWops* rw = SDL_RWFromMem (m_data, m_size);
-
-	//The above might fail an return null.
-	if (!rw) {
-		delete [] m_data;
-		m_data = 0;
-		PHYSFS_close(myfile);
-		cerr << "Error: Curropt data file!" << endl;
-		return nullptr;
-	}
-
-	SDL_Surface* surface = IMG_Load_RW(rw,true); //the second argument tells the function to three RWops
-
-	return surface;
-}
-
-shared_ptr<CppSdl::CppSdlImageHolder> IMG_Load3(string path) {
-	if (!PHYSFS_exists(path.c_str())) {
-		cerr << "Error: File not in blockattack.data: " << path << endl;
-		throw exception();
-	}
-
-	PHYSFS_file* myfile = PHYSFS_openRead(path.c_str());
-
-	// Get the lenght of the file
-	unsigned int m_size = PHYSFS_fileLength(myfile);
-
-	// Get the file data.
-	char* m_data = new char[m_size];
-
-	int length_read = PHYSFS_read (myfile, m_data, 1, m_size);
-
-	if (length_read != (int)m_size) {
-		delete [] m_data;
-		m_data = 0;
-		PHYSFS_close(myfile);
-		cerr << "Error: Curropt data file!" << endl;
-		throw exception();
-	}
-
-	PHYSFS_close(myfile);
-
-	shared_ptr<CppSdl::CppSdlImageHolder> surface(new CppSdl::CppSdlImageHolder(m_data, m_size));
-
-	return surface;
-}
-
 static void UnloadImages();
-static int InitImages();
+static int InitImages(sago::SagoSpriteHolder& holder);
 
 static string oldThemePath = "default";
 static bool loaded = false;
 
-void loadTheme(const string& themeName) {
+void loadTheme(sago::SagoSpriteHolder& holder, const string& themeName) {
 	if (loaded) {
 		UnloadImages();
 	}
@@ -186,7 +112,7 @@ void loadTheme(const string& themeName) {
 	//Look in blockattack.data
 	PHYSFS_addToSearchPath(((string)SHAREDIR+"/blockattack.data").c_str(), 1);
 	//Look in folder
-	PHYSFS_addToSearchPath(PHYSFS_getBaseDir(), 1);
+	PHYSFS_addToSearchPath( ((string) PHYSFS_getBaseDir()+"/data").c_str(), 1);
 	//Look in home folder
 #if defined(__unix__) || defined(_WIN32)
 	PHYSFS_addToSearchPath(home.c_str(), 1);
@@ -196,7 +122,7 @@ void loadTheme(const string& themeName) {
 		Config::getInstance()->setString("themename", themeName);
 	}
 	if (themeName.compare("default")==0 || (themeName.compare("start")==0)) {
-		InitImages();
+		InitImages(holder);
 		loaded =true;
 		return; //Nothing more to do
 	}
@@ -205,7 +131,7 @@ void loadTheme(const string& themeName) {
 #if defined(__unix__) || defined(_WIN32)
 	PHYSFS_addToSearchPath((home+(string)"/"+oldThemePath).c_str(), 0);
 #endif
-	InitImages();
+	InitImages(holder);
 	loaded = true;
 }
 
@@ -351,105 +277,75 @@ Mix_Chunk* Mix_LoadWAV2(const char* path) {
 }
 
 //Load all image files to memory
-static int InitImages() {
-	if (!((backgroundImage = IMG_Load2("gfx/background.png"))
-	        && (bOptions = IMG_Load2("gfx/bOptions.png"))
-	        && (bConfigure = IMG_Load2("gfx/bConfigure.png"))
-	        && (bSelectPuzzle = IMG_Load2("gfx/bSelectPuzzle.png"))
-	        && (bHighScore = IMG_Load2("gfx/bHighScore.png"))
-	        && (bBack = IMG_Load2("gfx/bBlank.png"))
-	        && (bForward = IMG_Load2("gfx/bForward.png"))
-	        && (blackLine = IMG_Load2("gfx/blackLine.png"))
-	        && (stageBobble = IMG_Load2("gfx/iStageClearLimit.png"))
-	        && (bricks[0] = IMG_Load2("gfx/bricks/blue.png"))
-	        && (bricks[1] = IMG_Load2("gfx/bricks/green.png"))
-	        && (bricks[2] = IMG_Load2("gfx/bricks/purple.png"))
-	        && (bricks[3] = IMG_Load2("gfx/bricks/red.png"))
-	        && (bricks[4] = IMG_Load2("gfx/bricks/turkish.png"))
-	        && (bricks[5] = IMG_Load2("gfx/bricks/yellow.png"))
-	        && (bricks[6] = IMG_Load2("gfx/bricks/grey.png"))
-	        && (crossover = IMG_Load2("gfx/crossover.png"))
-	        && (balls[0] = IMG_Load2("gfx/balls/ballBlue.png"))
-	        && (balls[1] = IMG_Load2("gfx/balls/ballGreen.png"))
-	        && (balls[2] = IMG_Load2("gfx/balls/ballPurple.png"))
-	        && (balls[3] = IMG_Load2("gfx/balls/ballRed.png"))
-	        && (balls[4] = IMG_Load2("gfx/balls/ballTurkish.png"))
-	        && (balls[5] = IMG_Load2("gfx/balls/ballYellow.png"))
-	        && (balls[6] = IMG_Load2("gfx/balls/ballGray.png"))
-	        && (cursor[0] = IMG_Load2("gfx/animations/cursor/1.png"))
-	        && (cursor[1] = IMG_Load2("gfx/animations/cursor/2.png"))
-	        && (bomb[0] = IMG_Load2("gfx/animations/bomb/bomb_1.png"))
-	        && (bomb[1] = IMG_Load2("gfx/animations/bomb/bomb_2.png"))
-	        && (ready[0] = IMG_Load2("gfx/animations/ready/ready_1.png"))
-	        && (ready[1] = IMG_Load2("gfx/animations/ready/ready_2.png"))
-	        && (explosion[0] = IMG_Load2("gfx/animations/explosion/0.png"))
-	        && (explosion[1] = IMG_Load2("gfx/animations/explosion/1.png"))
-	        && (explosion[2] = IMG_Load2("gfx/animations/explosion/2.png"))
-	        && (explosion[3] = IMG_Load2("gfx/animations/explosion/3.png"))
-	        && (counter[0] = IMG_Load2("gfx/counter/1.png"))
-	        && (counter[1] = IMG_Load2("gfx/counter/2.png"))
-	        && (counter[2] = IMG_Load2("gfx/counter/3.png"))
-	        && (backBoard = IMG_Load2("gfx/BackBoard.png")) //not used, we just test if it exists :)
-	        && (iGameOver = IMG_Load2("gfx/iGameOver.png"))
-	        && (iWinner = IMG_Load2("gfx/iWinner.png"))
-	        && (iDraw = IMG_Load2("gfx/iDraw.png"))
-	        && (iLoser = IMG_Load2("gfx/iLoser.png"))
-	        && (iChainBack = IMG_Load2("gfx/chainFrame.png"))
-//            && (optionsBack = IMG_Load2("gfx/options.png"))
-	        && (bOn = IMG_Load2("gfx/bOn.png"))
-	        && (bOff = IMG_Load2("gfx/bOff.png"))
-//            && (bChange = IMG_Load2("gfx/bChange.png"))
-	        && (b1024 = IMG_Load2("gfx/b1024.png"))
-	        && (dialogBox = IMG_Load2("gfx/dialogbox.png"))
-//	&& (fileDialogBox = IMG_Load2("gfx/fileDialogbox.png"))
-	        && (iLevelCheck = IMG_Load2("gfx/iLevelCheck.png"))
-	        && (iLevelCheckBox = IMG_Load2("gfx/iLevelCheckBox.png"))
-	        && (iLevelCheckBoxMarked = IMG_Load2("gfx/iLevelCheckBoxMarked.png"))
-	        && (iCheckBoxArea = IMG_Load2("gfx/iCheckBoxArea.png"))
-	        && (boardBackBack = IMG_Load2("gfx/boardBackBack.png"))
-	        && (changeButtonsBack = IMG_Load2("gfx/changeButtonsBack.png"))
-	        && (garbageTL = IMG_Load2("gfx/garbage/garbageTL.png"))
-	        && (garbageT = IMG_Load2("gfx/garbage/garbageT.png"))
-	        && (garbageTR = IMG_Load2("gfx/garbage/garbageTR.png"))
-	        && (garbageR = IMG_Load2("gfx/garbage/garbageR.png"))
-	        && (garbageBR = IMG_Load2("gfx/garbage/garbageBR.png"))
-	        && (garbageB = IMG_Load2("gfx/garbage/garbageB.png"))
-	        && (garbageBL = IMG_Load2("gfx/garbage/garbageBL.png"))
-	        && (garbageL = IMG_Load2("gfx/garbage/garbageL.png"))
-	        && (garbageFill = IMG_Load2("gfx/garbage/garbageFill.png"))
-	        && (garbageML = IMG_Load2("gfx/garbage/garbageML.png"))
-	        && (garbageM = IMG_Load2("gfx/garbage/garbageM.png"))
-	        && (garbageMR = IMG_Load2("gfx/garbage/garbageMR.png"))
-	        && (garbageGM = IMG_Load2("gfx/garbage/garbageGM.png"))
-	        && (garbageGML = IMG_Load2("gfx/garbage/garbageGML.png"))
-	        && (garbageGMR = IMG_Load2("gfx/garbage/garbageGMR.png"))
-	        && (smiley[0] = IMG_Load2("gfx/smileys/0.png"))
-	        && (smiley[1] = IMG_Load2("gfx/smileys/1.png"))
-	        && (smiley[2] = IMG_Load2("gfx/smileys/2.png"))
-	        && (smiley[3] = IMG_Load2("gfx/smileys/3.png"))
-	        //new in 1.3.2
-	        && (transCover = IMG_Load2("gfx/transCover.png"))
-	        //end new in 1.3.2
-	        //new in 1.4.0
-	        && (bSkip = IMG_Load2("gfx/bBlank.png"))
-	        && (bNext = IMG_Load2("gfx/bBlank.png"))
-	        && (bRetry = IMG_Load2("gfx/bBlank.png"))
-	     ))
-		//if there was a problem ie. "File not found"
-	{
-		cerr << "Error: Failed to load image file: " << SDL_GetError() << endl;
-		exit(1);
-	}
-	try {
-		bNewGame = IMG_Load3("gfx/bNewGame.png");
-		mouse = IMG_Load3("gfx/mouse.png");
-		menuMarked = IMG_Load3("gfx/menu/marked.png");
-		menuUnmarked = IMG_Load3("gfx/menu/unmarked.png");
-	}
-	catch (exception& e) {
-		cerr << e.what() << endl;
-		exit(1);
-	}
+static int InitImages(sago::SagoSpriteHolder& holder) {
+	bricks[0] = holder.GetSprite("blue");
+	bricks[1] = holder.GetSprite("green");
+	bricks[2] = holder.GetSprite("purple");
+	bricks[3] = holder.GetSprite("red");
+	bricks[4] = holder.GetSprite("turkish");
+	bricks[5] = holder.GetSprite("yellow");
+	bricks[6] = holder.GetSprite("grey");
+	bomb = holder.GetSprite("block_bomb");
+	backgroundImage = holder.GetSprite("background");
+	bHighScore = holder.GetSprite("b_highscore");
+	bBack = holder.GetSprite("b_back");
+	bForward = holder.GetSprite("b_forward");
+	blackLine = holder.GetSprite("black_line");
+	stageBobble = holder.GetSprite("i_stage_clear_limit");
+	crossover = holder.GetSprite("crossover");
+	balls[0] = holder.GetSprite("ball_blue");
+	balls[1] = holder.GetSprite("ball_green");
+	balls[2] = holder.GetSprite("ball_purple");
+	balls[3] = holder.GetSprite("ball_red");
+	balls[4] = holder.GetSprite("ball_turkish");
+	balls[5] = holder.GetSprite("ball_yellow");
+	balls[6] = holder.GetSprite("ball_gray");
+	cursor = holder.GetSprite("cursor");
+	ready = holder.GetSprite("ready");
+	explosion[0] = holder.GetSprite("explosion0");
+	explosion[2] = holder.GetSprite("explosion1");
+	explosion[3] = holder.GetSprite("explosion2");
+	explosion[4] = holder.GetSprite("explosion3");
+	counter[0] = holder.GetSprite("counter_1");
+	counter[1] = holder.GetSprite("counter_2");
+	counter[2] = holder.GetSprite("counter_3");
+	iGameOver = holder.GetSprite("i_game_over");
+	iWinner = holder.GetSprite("i_winner");
+	iDraw = holder.GetSprite("i_draw");
+	iLoser = holder.GetSprite("i_loser");
+	//iChainFrame = holder.GetSprite("chain_frame");
+	dialogBox = holder.GetSprite("dialogbox");
+	iLevelCheck = holder.GetSprite("i_level_check");
+	iLevelCheckBox = holder.GetSprite("i_level_check_box");
+	iLevelCheckBoxMarked = holder.GetSprite("i_level_check_box_marked");
+	iCheckBoxArea = holder.GetSprite("i_check_box_area");
+	boardBackBack = holder.GetSprite("board_back_back");
+	garbageTL = holder.GetSprite("garbage_tl");
+	garbageT = holder.GetSprite("garbage_t");
+	garbageTR = holder.GetSprite("garbage_tr");
+	garbageR = holder.GetSprite("garbage_r");
+	garbageBR = holder.GetSprite("garbage_br");
+	garbageB = holder.GetSprite("garbage_b");
+	garbageBL = holder.GetSprite("garbage_bl");
+	garbageL = holder.GetSprite("garbage_l");
+	garbageFill = holder.GetSprite("garbage_fill");
+	garbageML = holder.GetSprite("garbage_ml");
+	garbageM = holder.GetSprite("garbage_m");
+	garbageMR = holder.GetSprite("garbage_mr");
+	garbageGM = holder.GetSprite("garbage_gm");
+	garbageGML = holder.GetSprite("garbage_gml");
+	garbageGMR = holder.GetSprite("garbage_gmr");
+	smiley[0] = holder.GetSprite("smileys0");
+	smiley[1] = holder.GetSprite("smileys1");
+	smiley[2] = holder.GetSprite("smileys2");
+	smiley[3] = holder.GetSprite("smileys3");
+	transCover = holder.GetSprite("trans_cover");
+	bSkip = holder.GetSprite("b_skip");
+	bNext = holder.GetSprite("b_next");
+	bRetry = holder.GetSprite("b_retry");
+	mouse = holder.GetSprite("mouse");
+	menuMarked = holder.GetSprite("menu_marked");
+	menuUnmarked = holder.GetSprite("menu_unmarked");
 	
 	SDL_Color nf_button_color, nf_standard_blue_color, nf_standard_small_color;
 	memset(&nf_button_color,0,sizeof(SDL_Color));
@@ -491,83 +387,6 @@ void UnloadImages() {
 	if (verboseLevel) {
 		cout << "Unloading data..." << endl;
 	}
-	if (!NoSound) { //Only unload then it has been loaded!
-		Mix_HaltMusic();
-		Mix_FreeMusic(bgMusic);
-		Mix_FreeMusic(highbeatMusic);
-		Mix_FreeChunk(boing);
-		Mix_FreeChunk(applause);
-		Mix_FreeChunk(photoClick);
-		Mix_FreeChunk(counterChunk);
-		Mix_FreeChunk(counterFinalChunk);
-		Mix_FreeChunk(typingChunk);
-	}
-	//Free surfaces:
-	//I think this will crash, at least it happend to me...
-	//Chrashes no more. Caused by an undocumented double free
-	SDL_FreeSurface(backgroundImage);
-	SDL_FreeSurface(bOptions);
-	SDL_FreeSurface(bConfigure);
-	SDL_FreeSurface(bSelectPuzzle);
-	SDL_FreeSurface(bHighScore);
-	SDL_FreeSurface(blackLine);
-	SDL_FreeSurface(stageBobble);
-	SDL_FreeSurface(bricks[0]);
-	SDL_FreeSurface(bricks[1]);
-	SDL_FreeSurface(bricks[2]);
-	SDL_FreeSurface(bricks[3]);
-	SDL_FreeSurface(bricks[4]);
-	SDL_FreeSurface(bricks[5]);
-	SDL_FreeSurface(bricks[6]);
-	SDL_FreeSurface(crossover);
-	SDL_FreeSurface(balls[0]);
-	SDL_FreeSurface(balls[1]);
-	SDL_FreeSurface(balls[2]);
-	SDL_FreeSurface(balls[3]);
-	SDL_FreeSurface(balls[4]);
-	SDL_FreeSurface(balls[5]);
-	SDL_FreeSurface(balls[6]);
-	SDL_FreeSurface(cursor[0]);
-	SDL_FreeSurface(cursor[1]);
-	SDL_FreeSurface(backBoard); //not used, we just test if it exists :)
-	SDL_FreeSurface(iGameOver);
-	SDL_FreeSurface(iWinner);
-	SDL_FreeSurface(iDraw);
-	SDL_FreeSurface(iLoser);
-	SDL_FreeSurface(iChainBack);
-	SDL_FreeSurface(bOn);
-	SDL_FreeSurface(bOff);
-	SDL_FreeSurface(b1024);
-	SDL_FreeSurface(dialogBox);
-	//SDL_FreeSurface(fileDialogBox);
-	SDL_FreeSurface(iLevelCheck);
-	SDL_FreeSurface(iLevelCheckBox);
-	SDL_FreeSurface(iLevelCheckBoxMarked);
-	SDL_FreeSurface(iCheckBoxArea);
-	SDL_FreeSurface(boardBackBack);
-	SDL_FreeSurface(changeButtonsBack);
-	SDL_FreeSurface(garbageTL);
-	SDL_FreeSurface(garbageT);
-	SDL_FreeSurface(garbageTR);
-	SDL_FreeSurface(garbageR);
-	SDL_FreeSurface(garbageBR);
-	SDL_FreeSurface(garbageB);
-	SDL_FreeSurface(garbageBL);
-	SDL_FreeSurface(garbageL);
-	SDL_FreeSurface(garbageFill);
-	SDL_FreeSurface(garbageML);
-	SDL_FreeSurface(garbageM);
-	SDL_FreeSurface(garbageMR);
-	SDL_FreeSurface(garbageGML);
-	SDL_FreeSurface(garbageGM);
-	SDL_FreeSurface(garbageGMR);
-	SDL_FreeSurface(smiley[0]);
-	SDL_FreeSurface(smiley[1]);
-	SDL_FreeSurface(smiley[2]);
-	SDL_FreeSurface(smiley[3]);
-	SDL_FreeSurface(transCover);
-	mouse->MakeNull();
-	bNewGame->MakeNull();
 }
 
 
@@ -923,7 +742,7 @@ public:
 	}
 	
 	void DrawImgBoardBounded(sago::SagoSprite& img, int x, int y) const {
-		DrawIMG_Bounded(img, screen, x+topx, y+topy, topx, topy, topx + backBoard->w, topy + backBoard->h);
+		DrawIMG_Bounded(img, screen, x+topx, y+topy, topx, topy, topx + backBoard.GetWidth(), topy + backBoard.GetHeight());
 	}
 	
 	void PrintTextCenteredBoard(int x, int y, const char* text) {
@@ -994,10 +813,10 @@ private:
 				if ((board[j][i]%10 != -1) && (board[j][i]%10 < 7) && ((board[j][i]/1000000)%10==0)) {
 					DrawImgBoardBounded(bricks[board[j][i]%10],  j*bsize, bsize*12-i*bsize-pixels);
 					if ((board[j][i]/BLOCKWAIT)%10==1) {
-						DrawImgBoard(bomb[(ticks/BOMBTIME)%2],  j*bsize, bsize*12-i*bsize-pixels);
+						DrawImgBoard(bomb,  j*bsize, bsize*12-i*bsize-pixels);
 					}
 					if ((board[j][i]/BLOCKHANG)%10==1) {
-						DrawImgBoardBounded(ready[(ticks/READYTIME)%2],  j*bsize, bsize*12-i*bsize-pixels);
+						DrawImgBoardBounded(ready,  j*bsize, bsize*12-i*bsize-pixels);
 					}
 
 				}
@@ -1132,12 +951,11 @@ public:
 	void DoPaintJob() {
 		DrawIMG(boardBackBack,screen,this->GetTopX()-60,this->GetTopY()-68);
 		
-		nf_scoreboard_font.draw(this->GetTopX()+310,this->GetTopY()-68+148,_("Score:") );
-		nf_scoreboard_font.draw(this->GetTopX()+310,this->GetTopY()-68+197,_("Time:") );
-		nf_scoreboard_font.draw(this->GetTopX()+310,this->GetTopY()-68+246,_("Chain:") );
-		nf_scoreboard_font.draw(this->GetTopX()+310,this->GetTopY()-68+295,_("Speed:") );
+		nf_scoreboard_font.draw(screen, this->GetTopX()+310,this->GetTopY()-68+148,_("Score:") );
+		nf_scoreboard_font.draw(screen, this->GetTopX()+310,this->GetTopY()-68+197,_("Time:") );
+		nf_scoreboard_font.draw(screen, this->GetTopX()+310,this->GetTopY()-68+246,_("Chain:") );
+		nf_scoreboard_font.draw(screen, this->GetTopX()+310,this->GetTopY()-68+295,_("Speed:") );
 		DrawImgBoard(backBoard,  0, 0);
-		nf_standard_blue_font.setDest(screen); //reset to screen at the end of this funciton!
 
 		PaintBricks();
 		if (stageClear) {
@@ -1146,7 +964,7 @@ public:
 		if (puzzleMode&&(!bGameOver)) {
 			//We need to write nr. of moves left!
 			strHolder = "Moves left: " + itoa(MovesLeft);
-			nf_standard_blue_font.draw(topx+5, topy+5, "%s",strHolder.c_str());
+			nf_standard_blue_font.draw(screen, topx+5, topy+5, "%s",strHolder.c_str());
 
 		}
 		if (puzzleMode && stageButtonStatus == SBpuzzleMode) {
@@ -1164,7 +982,7 @@ public:
 			}
 			else {
 				strHolder = "Last puzzle";
-				nf_standard_blue_font.draw(topx+5, topy+5, "%s",strHolder.c_str());
+				nf_standard_blue_font.draw(screen, topx+5, topy+5, "%s",strHolder.c_str());
 			}
 		}
 		if (stageClear && stageButtonStatus == SBstageClear) {
@@ -1182,7 +1000,7 @@ public:
 			}
 			else {
 				strHolder = "Last stage";
-				nf_standard_blue_font.draw(topx+5, topy+5, "%s",strHolder.c_str());
+				nf_standard_blue_font.draw(screen, topx+5, topy+5, "%s",strHolder.c_str());
 			}
 		}
 
@@ -1190,11 +1008,11 @@ public:
 		if (AI_Enabled&&(!bGameOver)) {
 			strHolder = "AI_status: " + itoa(AIstatus)+ ", "+ itoa(AIlineToClear);
 			//NFont_Write(   5, 5, strHolder.c_str());
-			nf_standard_blue_font.draw(topx+5, topy+5, "%s",strHolder.c_str());
+			nf_standard_blue_font.draw(screen, topx+5, topy+5, "%s",strHolder.c_str());
 		}
 #endif
 		if (!bGameOver) {
-			DrawImgBoard(cursor[(ticks/600)%2],cursorx*bsize-4,11*bsize-cursory*bsize-pixels-4);
+			DrawImgBoard(cursor,cursorx*bsize-4,11*bsize-cursory*bsize-pixels-4);
 		}
 		if (ticks<gameStartedAt) {
 			int currentCounter = abs((int)ticks-(int)gameStartedAt)/1000;
@@ -1247,7 +1065,6 @@ public:
 				}
 			}
 		}
-		nf_standard_blue_font.setDest(screen);
 	}
 
 
@@ -1285,15 +1102,17 @@ void writeScreenShot() {
 #else
 	string buf = "screenshot"+itoa(rightNow)+".bmp";
 #endif
-	SDL_SaveBMP( screen, buf.c_str() );
-	if (!NoSound)
+	//SDL_SaveBMP( screen, buf.c_str() );
+	//TODO: Write screenshot
+	if (!NoSound) {
 		if (SoundEnabled) {
 			Mix_PlayChannel(1,photoClick,0);
 		}
+	}
 }
 
 //Function to return the name of a key, to be displayed...
-static string getKeyName(SDLKey key) {
+static string getKeyName(SDL_Keycode key) {
 	string keyname(SDL_GetKeyName(key));
 	return keyname;
 }
@@ -1306,7 +1125,6 @@ bool OpenDialogbox(int x, int y, std::string& name) {
 	ReadKeyboard rk = ReadKeyboard(name.c_str());
 	string strHolder;
 	DrawIMG(backgroundImage,screen,0,0);
-	SDLUnicodeScope unicodeScope;
 	while (!done && !Config::getInstance()->isShuttingDown()) {
 		DrawIMG(dialogBox,screen,x,y);
 		NFont_Write(screen, x+40,y+76,rk.GetString());
@@ -1336,12 +1154,12 @@ bool OpenDialogbox(int x, int y, std::string& name) {
 					accept = false;
 				}
 				else if (!(event.key.keysym.sym == SDLK_BACKSPACE)) {
-					if ((rk.ReadKey(event.key.keysym))&&(SoundEnabled)&&(!NoSound)) {
+					if ((rk.ReadKey(event))&&(SoundEnabled)&&(!NoSound)) {
 						Mix_PlayChannel(1,typingChunk,0);
 					}
 				}
 				else if (event.key.keysym.sym == SDLK_BACKSPACE) {
-					if ((rk.ReadKey(event.key.keysym))&&(SoundEnabled)&&(!NoSound)) {
+					if ((rk.ReadKey(event))&&(SoundEnabled)&&(!NoSound)) {
 						Mix_PlayChannel(1,typingChunk,0);
 					}
 				}
@@ -1349,12 +1167,11 @@ bool OpenDialogbox(int x, int y, std::string& name) {
 
 		}   //while(event)
 
-		SDL_Flip(screen); //Update screen
+		SDL_RenderPresent(screen); //Update screen
 	}   //while(!done)
 	name = rk.GetString();
 	bScreenLocked = false;
 	showDialog = false;
-	unicodeScope.Release();
 	return accept;
 }
 
@@ -1362,10 +1179,10 @@ bool OpenDialogbox(int x, int y, std::string& name) {
 void DrawHighscores(int x, int y, bool endless) {
 	DrawIMG(backgroundImage,screen,0,0);
 	if (endless) {
-		nf_standard_blue_font.draw(x+100,y+100, "%s",_("Endless:") );
+		nf_standard_blue_font.draw(screen, x+100,y+100, "%s",_("Endless:") );
 	}
 	else {
-		nf_standard_blue_font.draw(x+100,y+100, "%s",_("Time Trial:") );
+		nf_standard_blue_font.draw(screen, x+100,y+100, "%s",_("Time Trial:") );
 	}
 	for (int i =0; i<10; i++) {
 		char playerScore[32];
@@ -1382,8 +1199,8 @@ void DrawHighscores(int x, int y, bool endless) {
 		else {
 			strcpy(playerName,theTopScoresTimeTrial.getScoreName(i));
 		}
-		nf_standard_blue_font.draw(x+420,y+150+i*35, "%s",playerScore);
-		nf_standard_blue_font.draw(x+60,y+150+i*35, "%s",playerName);
+		nf_standard_blue_font.draw(screen, x+420,y+150+i*35, "%s",playerScore);
+		nf_standard_blue_font.draw(screen, x+60,y+150+i*35, "%s",playerName);
 	}
 }
 
@@ -1478,9 +1295,9 @@ void OpenScoresDisplay() {
 		//Draw buttons:
 		DrawIMG(bHighScore,screen,scoreX,scoreY);
 		DrawIMG(bBack,screen,backX,backY);
-		nf_button_font.drawCenter(backX+60,backY+10,_("Back"));
+		nf_button_font.draw(screen, backX+60,backY+10, NFont::CENTER ,_("Back"));
 		DrawIMG(bNext,screen,nextX,nextY);
-		nf_button_font.drawCenter(nextX+60,nextY+10,_("Next"));
+		nf_button_font.draw(screen, nextX+60,nextY+10, NFont::CENTER,_("Next"));
 
 		//Draw page number
 		string pageXofY = (boost::format(_("Page %1% of %2%") )%(page+1)%numberOfPages).str();
@@ -1560,9 +1377,8 @@ void OpenScoresDisplay() {
 			}
 		}
 
-		//DrawIMG(mouse,screen,mousex,mousey);
-		mouse->PaintTo(screen,mousex,mousey);
-		SDL_Flip(screen); //Update screen
+		mouse.Draw(screen, SDL_GetTicks(), mousex, mousey);
+		SDL_RenderPresent(screen);
 	}
 
 
@@ -1586,11 +1402,10 @@ bool OpenFileDialogbox(int x, int y, char* name) {
 	while (!done && !Config::getInstance()->isShuttingDown()) {
 		DrawIMG(backgroundImage,screen,0,0);
 		DrawIMG(bForward,screen,x+460,y+420);
-		nf_button_font.drawCenter(x+20+60, y+420+10, _("Forward"));
+		nf_button_font.draw(screen, x+20+60, y+420+10, NFont::CENTER, _("Forward"));
 		DrawIMG(bBack,screen,x+20,y+420);
-		nf_button_font.drawCenter(x+20+60, y+420+10, _("Back"));
+		nf_button_font.draw(screen, x+20+60, y+420+10, NFont::CENTER, _("Back"));
 		const int nrOfFiles = 10;
-		DrawIMG(changeButtonsBack,screen,x,y);
 		for (int i=0; i<nrOfFiles; i++) {
 			NFont_Write(screen, x+10,y+10+36*i,lf.getFileName(i).c_str());
 		}
@@ -1649,9 +1464,8 @@ bool OpenFileDialogbox(int x, int y, char* name) {
 			}
 		}
 
-		//DrawIMG(mouse,screen,mousex,mousey);
-		mouse->PaintTo(screen,mousex,mousey);
-		SDL_Flip(screen); //Update screen
+		mouse.Draw(screen, SDL_GetTicks(), mousex, mousey);
+		SDL_RenderPresent(screen); //Update screen
 	}
 	return true;
 }
@@ -1670,8 +1484,9 @@ static void DrawBalls() {
 			//cout << "Printing text: " << theTextManeger.textArray[i].getText() << endl;
 			int x = theTextManeger.textArray[i].getX()-12;
 			int y = theTextManeger.textArray[i].getY()-12;
-			DrawIMG(iChainBack,screen,x,y);
-			nf_standard_small_font.drawCenter(x+12,y+7, "%s",theTextManeger.textArray[i].getText());
+			//DrawIMG(iChainBack,screen,x,y);
+			
+			nf_standard_small_font.draw(screen, x+12,y+7, NFont::CENTER, "%s",theTextManeger.textArray[i].getText());
 		}
 	} //for
 }    //DrawBalls
@@ -1871,10 +1686,10 @@ void DrawEverything(int xsize, int ysize,BlockGameSdl* theGame, BlockGameSdl* th
 	}
 
 	//NFont_Write(screen, 800,4,FPS);
-	nf_standard_blue_font.draw(800, 4, "%s", FPS);
+	nf_standard_blue_font.draw(screen, 800, 4, "%s", FPS);
 #endif
 
-	//SDL_Flip(screen); Update screen is now called outside DrawEvrything, bacause the mouse needs to be painted
+	//SDL_RenderPresent(screen); Update screen is now called outside DrawEvrything, bacause the mouse needs to be painted
 
 }
 
@@ -2011,7 +1826,7 @@ int PuzzleLevelSelect(int Type) {
 				}
 			}
 
-		SDL_GetKeyState(nullptr);
+		SDL_GetKeyboardState(nullptr);
 
 		SDL_GetMouseState(&mousex,&mousey);
 		if (mousex != oldmousex || mousey != oldmousey) {
@@ -2072,9 +1887,8 @@ int PuzzleLevelSelect(int Type) {
 			NFont_Write(screen, 200,600,totalString.c_str());
 		}
 
-		//DrawIMG(mouse,screen,mousex,mousey);
-		mouse->PaintTo(screen,mousex,mousey);
-		SDL_Flip(screen); //draws it all to the screen
+		mouse.Draw(screen, SDL_GetTicks(), mousex, mousey);
+		SDL_RenderPresent(screen); //draws it all to the screen
 
 	}
 	DrawIMG(backgroundImage, screen, 0, 0);
@@ -2416,41 +2230,41 @@ int main(int argc, char* argv[]) {
 		joyplay2 = (bool)configSettings->getInt("joypad2");
 
 		if (configSettings->exists("player1keyup")) {
-			keySettings[0].up = (SDLKey)configSettings->getInt("player1keyup");
+			keySettings[0].up = (SDL_Keycode)configSettings->getInt("player1keyup");
 		}
 		if (configSettings->exists("player1keydown")) {
-			keySettings[0].down = (SDLKey)configSettings->getInt("player1keydown");
+			keySettings[0].down = (SDL_Keycode)configSettings->getInt("player1keydown");
 		}
 		if (configSettings->exists("player1keyleft")) {
-			keySettings[0].left = (SDLKey)configSettings->getInt("player1keyleft");
+			keySettings[0].left = (SDL_Keycode)configSettings->getInt("player1keyleft");
 		}
 		if (configSettings->exists("player1keyright")) {
-			keySettings[0].right = (SDLKey)configSettings->getInt("player1keyright");
+			keySettings[0].right = (SDL_Keycode)configSettings->getInt("player1keyright");
 		}
 		if (configSettings->exists("player1keychange")) {
-			keySettings[0].change = (SDLKey)configSettings->getInt("player1keychange");
+			keySettings[0].change = (SDL_Keycode)configSettings->getInt("player1keychange");
 		}
 		if (configSettings->exists("player1keypush")) {
-			keySettings[0].push = (SDLKey)configSettings->getInt("player1keypush");
+			keySettings[0].push = (SDL_Keycode)configSettings->getInt("player1keypush");
 		}
 
 		if (configSettings->exists("player2keyup")) {
-			keySettings[2].up = (SDLKey)configSettings->getInt("player2keyup");
+			keySettings[2].up = (SDL_Keycode)configSettings->getInt("player2keyup");
 		}
 		if (configSettings->exists("player2keydown")) {
-			keySettings[2].down = (SDLKey)configSettings->getInt("player2keydown");
+			keySettings[2].down = (SDL_Keycode)configSettings->getInt("player2keydown");
 		}
 		if (configSettings->exists("player2keyleft")) {
-			keySettings[2].left = (SDLKey)configSettings->getInt("player2keyleft");
+			keySettings[2].left = (SDL_Keycode)configSettings->getInt("player2keyleft");
 		}
 		if (configSettings->exists("player2keyright")) {
-			keySettings[2].right = (SDLKey)configSettings->getInt("player2keyright");
+			keySettings[2].right = (SDL_Keycode)configSettings->getInt("player2keyright");
 		}
 		if (configSettings->exists("player2keychange")) {
-			keySettings[2].change = (SDLKey)configSettings->getInt("player2keychange");
+			keySettings[2].change = (SDL_Keycode)configSettings->getInt("player2keychange");
 		}
 		if (configSettings->exists("player2keypush")) {
-			keySettings[2].push = (SDLKey)configSettings->getInt("player2keypush");
+			keySettings[2].push = (SDL_Keycode)configSettings->getInt("player2keypush");
 		}
 		if (configSettings->exists("player1name")) {
 			player1name = configSettings->getString("player1name");
@@ -2501,11 +2315,14 @@ int main(int argc, char* argv[]) {
 
 	//Init the file system abstraction layer
 	PHYSFS_init(argv[0]);
+	PHYSFS_addToSearchPath( ((string) PHYSFS_getBaseDir()+"/data").c_str(), 1);
 	//Load default theme
-	loadTheme(Config::getInstance()->getString("themename"));
+	sago::SagoDataHolder d(renderer);
+	sago::SagoSpriteHolder spriteholder(d);
+	loadTheme(spriteholder, Config::getInstance()->getString("themename"));
 	//Now sets the icon:
-	SDL_Surface* icon = IMG_Load2("gfx/icon.png");
-	SDL_WM_SetIcon(icon,nullptr);
+	//SDL_Surface* icon = IMG_Load2("gfx/icon.png");
+	//SDL_WM_SetIcon(icon,nullptr);
 
 	if (verboseLevel) {
 		cout << "Images loaded" << endl;
@@ -2538,7 +2355,7 @@ int main(int argc, char* argv[]) {
 	}
 	DrawIMG(backgroundImage, screen, 0, 0);
 	DrawEverything(xsize,ysize,&theGame,&theGame2);
-	SDL_Flip(screen);
+	SDL_RenderPresent(screen);
 	//game loop
 	MainMenu();
 
@@ -2601,7 +2418,7 @@ int main(int argc, char* argv[]) {
 
 
 int runGame(int gametype, int level) {
-	Uint8* keys;
+	const Uint8* keys;
 	int mousex, mousey;   //Mouse coordinates
 	bScreenLocked = false;
 	theTopScoresEndless = Highscore(1);
@@ -2723,7 +2540,7 @@ int runGame(int gametype, int level) {
 			mustsetupgame = false;
 			DrawIMG(backgroundImage, screen, 0, 0);
 			DrawEverything(xsize,ysize,&theGame,&theGame2);
-			SDL_Flip(screen);
+			SDL_RenderPresent(screen);
 		}
 
 		if (!(highPriority)) {
@@ -2851,7 +2668,7 @@ int runGame(int gametype, int level) {
 			**************************** Repeating start **************************
 			**********************************************************************/
 
-			keys = SDL_GetKeyState(nullptr);
+			keys = SDL_GetKeyboardState(nullptr);
 //Also the joysticks:
 //Repeating not implemented
 
@@ -3078,7 +2895,7 @@ int runGame(int gametype, int level) {
 			**********************************************************************/
 
 
-			SDL_GetKeyState(nullptr);
+			SDL_GetKeyboardState(nullptr);
 
 			SDL_GetMouseState(&mousex,&mousey);
 
@@ -3301,8 +3118,8 @@ int runGame(int gametype, int level) {
 		oldMousex = mousex;
 		oldMousey = mousey;
 		//Draw the mouse:
-		mouse->PaintTo(screen,mousex,mousey);
-		SDL_Flip(screen);
+		mouse.Draw(screen, SDL_GetTicks(), mousex, mousey);
+		SDL_RenderPresent(screen);
 	} //game loop
 	return 0;
 }
